Start implementing the get_spacing function
